Move executeSqlScript to SqlScriptingExecution and fix empty results
